# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fit Club client portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Appointment {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Start time (ISO 8601) */
  time: string;
  /** End time (ISO 8601) */
  endTime: string;
  /** Duration in minutes */
  duration: number;
  /** Appointment type name */
  type: string;
  /**
     * Calendar name
     * @nullable
     */
  calendar?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  /**
     * URL to Acuity confirmation page
     * @nullable
     */
  confirmationPage?: string | null;
}

export interface AppointmentSummary {
  upcomingCount: number;
  pastCount: number;
  nextAppointment?: Appointment | null;
}