# lib/api-zod/src/generated/types/appointmentSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fit Club client portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';

export interface AppointmentSummary {
  upcomingCount: number;
  pastCount: number;
  nextAppointment?: Appointment | null;
}